+++ pyccel/codegen/printing/pycode.py
@@ -3,18 +3,19 @@
 # This file is part of Pyccel which is released under MIT License. See the LICENSE file or #
 # go to https://github.com/pyccel/pyccel/blob/master/LICENSE for full license details.     #
 #------------------------------------------------------------------------------------------#
+import warnings
 
 from pyccel.decorators import __all__ as pyccel_decorators
 
-from pyccel.ast.builtins   import PythonMin, PythonMax
+from pyccel.ast.builtins   import PythonMin, PythonMax, PythonType
 from pyccel.ast.core       import CodeBlock, Import, Assign, FunctionCall, For, AsName, FunctionAddress
-from pyccel.ast.core       import IfSection, FunctionDef, Module, DottedFunctionCall
-from pyccel.ast.datatypes  import default_precision
+from pyccel.ast.core       import IfSection, FunctionDef, Module, DottedFunctionCall, PyccelFunctionDef
+from pyccel.ast.datatypes  import default_precision, datatype
 from pyccel.ast.functionalexpr import FunctionalFor
 from pyccel.ast.literals   import LiteralTrue, LiteralString
 from pyccel.ast.literals   import LiteralInteger, LiteralFloat, LiteralComplex
 from pyccel.ast.numpyext   import NumpyShape, NumpySize, numpy_target_swap
-from pyccel.ast.numpyext   import NumpyArray, NumpyNonZero
+from pyccel.ast.numpyext   import NumpyArray, NumpyNonZero, NumpyResultType
 from pyccel.ast.numpyext   import DtypePrecisionToCastFunction
 from pyccel.ast.variable   import DottedName, HomogeneousTupleVariable, Variable
 from pyccel.ast.utilities  import builtin_import_registry as pyccel_builtin_import_registry
@@ -151,7 +152,72 @@ class PythonCodePrinter(CodePrinter):
                 raise NotImplementedError("Type {} not handled in a FunctionalFor".format(type(body)))
         return body, iterables
 
+    def _get_numpy_name(self, expr):
+        """
+        Get the name of a NumPy function and ensure it is imported.
+
+        Get the name of a NumPy function from an instance of the class. The
+        name is saved in the class by default, however _aliases are checked
+        in case the function was imported explicitly with a different name
+        (e.g. `from numpy import int32 as i32`). If the name is not found in
+        aliases then it is added to the objects imported from NumPy.
+
+        Parameters
+        ----------
+        expr : PyccelInternalFunction
+            A pyccel node describing a NumPy function.
+
+        Returns
+        -------
+        str
+            The name that should be used in the code.
+        """
+        if isinstance(expr, type):
+            cls = expr
+        else:
+            cls = type(expr)
+        type_name = expr.name
+        name = self._aliases.get(cls, type_name)
+        if name == type_name:
+            self.insert_new_import(
+                    source = 'numpy',
+                    target = AsName(cls, name))
+        return name
+
     #----------------------------------------------------------------------
+
+    def _print_dtype_argument(self, expr, init_dtype):
+        """
+        Print a dtype argument.
+
+        Print the argument `dtype=X` from the dtype initially provided.
+
+        Parameters
+        ----------
+        expr : PyccelAstNode
+            The expression whose datatype is being determined.
+
+        init_dtype : PythonType, PyccelFunctionDef, LiteralString, str
+            The actual dtype passed to the NumPy function.
+
+        Returns
+        -------
+        str
+            The code for the dtype argument.
+        """
+        if init_dtype is None:
+            return ''
+        elif isinstance(init_dtype, (PythonType, NumpyResultType)):
+            dtype = self._print(init_dtype)
+            return "dtype = " + dtype
+        elif isinstance(init_dtype, PyccelFunctionDef):
+            dtype = self._get_numpy_name(init_dtype.cls_name)
+            return f"dtype={dtype}"
+        else:
+            dtype = self._print(expr.dtype)
+            if expr.precision != -1:
+                dtype = self._get_numpy_name(DtypePrecisionToCastFunction[datatype(dtype).name][expr.precision])
+            return f"dtype={dtype}"
 
     def _print_Header(self, expr):
         return ''
@@ -229,7 +295,17 @@ class PythonCodePrinter(CodePrinter):
         func = expr.functions[0]
         if not isinstance(func, FunctionAddress):
             func.rename(expr.name)
-        return self._print(func)
+        func_code = self._print(func)
+        _, body = func_code.split(':\n',1)
+        for func in expr.functions[1:]:
+            if not isinstance(func, FunctionAddress):
+                func.rename(expr.name)
+            i_func_code = self._print(func)
+            _, i_body = i_func_code.split(':\n',1)
+            if i_body != body:
+                warnings.warn(UserWarning("Generated code varies between interfaces but has not been printed. This Python code may produce unexpected results."))
+                return func_code
+        return func_code
 
     def _print_FunctionDef(self, expr):
         self.set_scope(expr.scope)
@@ -256,13 +332,13 @@ class PythonCodePrinter(CodePrinter):
                         name=name,
                         args=args,
                         body=body)
-        decorators = expr.decorators
+        decorators = expr.decorators.copy()
         if decorators:
             if decorators['template']:
                 # Eliminate template_dict because it is useless in the printing
-                expr.decorators['template'] = expr.decorators['template']['decorator_list']
+                decorators['template'] = decorators['template']['decorator_list']
             else:
-                expr.decorators.pop('template')
+                decorators.pop('template')
             for n,f in decorators.items():
                 if n in pyccel_decorators:
                     self.insert_new_import(DottedName('pyccel.decorators'), AsName(decorators_mod[n], n))
@@ -294,6 +370,13 @@ class PythonCodePrinter(CodePrinter):
 
         self.exit_scope()
         return code
+
+    def _print_PyccelFunctionDef(self, expr):
+        cls = expr.cls_name
+        if cls.__name__.startswith('Numpy'):
+            return self._get_numpy_name(cls)
+        else:
+            return cls.name
 
     def _print_FunctionAddress(self, expr):
         return expr.name
@@ -351,25 +434,13 @@ class PythonCodePrinter(CodePrinter):
     def _print_PythonInt(self, expr):
         name = 'int'
         if expr.precision != -1:
-            type_name = name + str(expr.precision*8)
-            cls       = type(expr)
-            name = self._aliases.get(cls, type_name)
-            if name == type_name:
-                self.insert_new_import(
-                        source = 'numpy',
-                        target = AsName(cls, name))
+            name = self._get_numpy_name(expr)
         return '{}({})'.format(name, self._print(expr.arg))
 
     def _print_PythonFloat(self, expr):
         name = 'float'
         if expr.precision != -1:
-            type_name = name + str(expr.precision*8)
-            cls       = type(expr)
-            name = self._aliases.get(cls, type_name)
-            if name == type_name:
-                self.insert_new_import(
-                        source = 'numpy',
-                        target = AsName(cls, name))
+            name = self._get_numpy_name(expr)
         return '{}({})'.format(name, self._print(expr.arg))
 
     def _print_PythonComplex(self, expr):
@@ -381,12 +452,7 @@ class PythonCodePrinter(CodePrinter):
 
     def _print_NumpyComplex(self, expr):
         if expr.precision != -1:
-            cls       = type(expr)
-            name = self._aliases.get(cls, expr.name)
-            if name == expr.name:
-                self.insert_new_import(
-                        source = 'numpy',
-                        target = AsName(cls, expr.name))
+            name = self._get_numpy_name(expr)
         else:
             name = 'complex'
         if expr.is_cast:
@@ -432,22 +498,12 @@ class PythonCodePrinter(CodePrinter):
     def _print_PyccelArrayShapeElement(self, expr):
         arg = self._print(expr.arg)
         index = self._print(expr.index)
-        expected_name = NumpyShape.name
-        name = self._aliases.get(NumpyShape, expected_name)
-        if name == expected_name:
-            self.insert_new_import(
-                    source = 'numpy',
-                    target = AsName(type(expr), expected_name))
+        name = self._get_numpy_name(expr)
         return f'{name}({arg})[{index}]'
 
     def _print_PyccelArraySize(self, expr):
         arg = self._print(expr.arg)
-        expected_name = NumpySize.name
-        name = self._aliases.get(NumpySize, expected_name)
-        if name == expected_name:
-            self.insert_new_import(
-                    source = 'numpy',
-                    target = AsName(type(expr), expected_name))
+        name = self._get_numpy_name(expr)
         return f'{name}({arg})'
 
     def _print_Comment(self, expr):
@@ -635,11 +691,10 @@ class PythonCodePrinter(CodePrinter):
         return'{0} {1}= {2}\n'.format(lhs,op,rhs)
 
     def _print_PythonRange(self, expr):
-        name = self._aliases.get(type(expr), expr.name)
         start = self._print(expr.start)
         stop  = self._print(expr.stop)
         step  = self._print(expr.step)
-        return '{}({}, {}, {})'.format(name,start,stop,step)
+        return f'range({start}, {stop}, {step})'
 
     def _print_Allocate(self, expr):
         return ''
@@ -648,49 +703,32 @@ class PythonCodePrinter(CodePrinter):
         return ''
 
     def _print_NumpyArray(self, expr):
-        dtype = self._print(expr.dtype)
-        if expr.precision != default_precision[str(expr.dtype)]:
-            factor = 16 if dtype == 'complex' else 8
-            dtype += str(expr.precision*factor)
+        name = self._get_numpy_name(expr)
 
-        name  = self._aliases.get(type(expr), expr.name)
         arg   = self._print(expr.arg)
-        dtype = "dtype={}".format(dtype)
-        order = "order='{}'".format(expr.order) if expr.order else ''
-        args  = [arg, dtype, order]
-        return "{name}({args})".format(
-                name  = name,
-                args  = ', '.join(a for a in args if a))
+        dtype = self._print_dtype_argument(expr, expr.init_dtype)
+        order = f"order='{expr.order}'" if expr.order else ''
+        args  = ', '.join(a for a in [arg, dtype, order] if a!= '')
+        return f"{name}({args})"
 
     def _print_NumpyAutoFill(self, expr):
         func_name = self._aliases.get(type(expr), expr.name)
 
-        dtype = self._print(expr.dtype)
-        if expr.precision != default_precision[str(expr.dtype)]:
-            factor = 16 if dtype == 'complex' else 8
-            dtype += str(expr.precision*factor)
-
+        dtype = self._print_dtype_argument(expr, expr.init_dtype)
         shape = self._print(expr.shape)
-        dtype = "dtype={}".format(dtype)
-        order = "order='{}'".format(expr.order) if expr.order else ''
-        args  = [shape, dtype, order]
-        return "{func_name}({args})".format(
-                func_name = func_name,
-                args  = ', '.join(a for a in args if a))
+        order = f"order='{expr.order}'" if expr.order else ''
+        args  = ', '.join(a for a in [shape, dtype, order] if a!='')
+        return f"{func_name}({args})"
 
     def _print_NumpyLinspace(self, expr):
         name = self._aliases.get(type(expr), expr.name)
-        dtype = self._print(expr.dtype)
-        factor = 16 if dtype == 'complex' else 8
-        dtype += str(expr.precision*factor)
-
-        return "{0}({1}, {2}, num={3}, endpoint={4}, dtype='{5}')".format(
-                name,
-                self._print(expr.start),
-                self._print(expr.stop),
-                self._print(expr.num),
-                self._print(expr.endpoint),
-                dtype)
+        dtype = self._print_dtype_argument(expr, expr.init_dtype)
+        start = self._print(expr.start)
+        stop = self._print(expr.stop)
+        num = "num = " + self._print(expr.num)
+        endpoint = "endpoint = "+self._print(expr.endpoint)
+        args = ', '.join(a for a in [start, stop, num, endpoint, dtype] if a != '')
+        return f"{name}({args})"
 
     def _print_NumpyMatmul(self, expr):
         name = self._aliases.get(type(expr), expr.name)
@@ -702,42 +740,38 @@ class PythonCodePrinter(CodePrinter):
 
     def _print_NumpyFull(self, expr):
         name = self._aliases.get(type(expr), expr.name)
-        dtype = self._print(expr.dtype)
-        if expr.precision != default_precision[str(expr.dtype)]:
-            factor = 16 if dtype == 'complex' else 8
-            dtype += str(expr.precision*factor)
 
+        dtype = self._print_dtype_argument(expr, expr.init_dtype)
         shape      = self._print(expr.shape)
         fill_value = self._print(expr.fill_value)
-        dtype      = "dtype={}".format(dtype)
-        order      = "order='{}'".format(expr.order) if expr.order else ''
-        args       = [shape, fill_value, dtype, order]
-        return "{name}({args})".format(
-                name  = name,
-                args  = ', '.join(a for a in args if a))
+        order      = f"order='{expr.order}'" if expr.order else ''
+        args       = ', '.join(a for a in [shape, fill_value, dtype, order] if a)
+        return f"{name}({args})"
 
     def _print_NumpyArange(self, expr):
         name = self._aliases.get(type(expr), expr.name)
-        return "{name}({start}, {stop}, {step}, dtype={dtype})".format(
-                name  = name,
-                start = self._print(expr.start),
-                stop  = self._print(expr.stop),
-                step  = self._print(expr.step),
-                dtype = self._print(expr.dtype))
+        dtype = self._print_dtype_argument(expr, expr.init_dtype)
+        args = ', '.join(a for a in [self._print(expr.start),
+                          self._print(expr.stop),
+                          self._print(expr.step),
+                          dtype] if a != '')
+        return f"{name}({args})"
 
     def _print_PyccelInternalFunction(self, expr):
         name = self._aliases.get(type(expr),expr.name)
         args = ', '.join(self._print(a) for a in expr.args)
         return "{}({})".format(name, args)
 
-    def _print_NumpyArray(self, expr):
-        name = self._aliases.get(type(expr),'array')
-        if name == 'array':
-            self.insert_new_import(
-                    source = 'numpy',
-                    target = AsName(NumpyArray, 'array'))
-        arg = self._print(expr.arg)
-        return "{}({})".format(name, arg)
+    def _print_NumpyResultType(self, expr):
+        args = expr.args
+        if len(args) == 1 and args[0].rank > 1:
+            arg = args[0]
+            arg_code = self._print(arg)
+            return f"{arg_code}.dtype"
+        else:
+            name = self._get_numpy_name(expr)
+            args = ', '.join(self._print(a) for a in expr.args)
+            return f"{name}({args})"
 
     def _print_NumpyRandint(self, expr):
         name = self._aliases.get(type(expr), expr.name)
@@ -1015,7 +1049,7 @@ class PythonCodePrinter(CodePrinter):
 
     def _print_PythonType(self, expr):
         return 'type({})'.format(self._print(expr.arg))
-    
+
     #-----------------Class Printer---------------------------------
 
     def _print_ClassDef(self, expr):