+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 -   Allow interfaces in classes.
 -   Python support for a simple class.
 -   #1430 : Added conjugate support to booleans.
+-   #1260 : Add support for NumPy `dtype` property: `a.dtype`.
+-   #1260 : Add support for NumPy `result_type` function.
 
 ### Fixed
 
@@ -26,6 +28,7 @@ All notable changes to this project will be documented in this file.
 
 ### Deprecated
 
+-   Stop generating `numpy.bool` (deprecated from NumPy) in code.
 -   \[INTERNALS\] Removed `obsolete` folder.
 -   \[INTERNALS\] Removed out of date `samples` folder.
 -   \[INTERNALS\] Removed out of date `doc` folder.